Remove range-metadata amdgpu workaround
Range metadata was disabled for amdgpu due to a backend bug. I did not
encounter any problems when removing the workaround to enable range
metadata (tried compiling `core` and `alloc`), so I suppose this has
been fixed in the backend in the last years.

Remove the workaround to re-enable range metadata.