Fix linker-plugin-lto only doing thin lto
When rust provides LLVM bitcode files to lld and the bitcode contains
function summaries as used for thin lto, lld defaults to using thin lto.
This prevents some optimizations that are only applied for fat lto. I
ran into this with the amdgpu backend and was able to create a test
there. Unfortunately, I wasn’t able to recreate the same test on x86,
the difference between thin and fat lto seems less pronounced there.

The important part of the change is setting the `ThinLTO=0` module flag.
The rest of the changes are fixing the `<function>.kd` symbol getting
exported for amdhsa with linker-plugin-lto.