Do not run mir opts for GPUs
GPU targets have convergent operations that require careful handling
when running optimizations. E.g. they must not be duplicated.

An example convergent operation is a barrier/syncthreads.

We do not want to deal with convergent operations in mir optimizations,
so set the optimization level to 0 and skip all optimizations.

This affects the amdgpu and nvptx targets.