Add intrinsic for dynamic group-shared memory on GPUs
Group-shared memory is a memory region that is shared between all
threads in a work-group on GPUs. Dynamic group-shared memory is in that
memory region, though the allocated size is specified late, when
launching a kernel, instead of early at compile-time.

With this change, dynamic group-shared memory can be accessed in Rust by
calling the new `gpu_dynamic_groupshared_mem<T>() -> *mut T` intrinsic.

It returns the pointer to dynamic group-shared memory guaranteeing that
it is aligned to at least the alignment of `T`.
The pointer is dereferencable for the size specified when launching the
current gpu-kernel (which may be the size of `T` but can also be larger
or smaller or zero).

All calls to this intrinsic return a pointer to the same address.

See the intrinsic documentation for more details.

It was also considered to expose dynamic group-shared memory as extern
static variables in Rust, like they are represented in LLVM IR. However,
due to the pointer not being guaranteed to be dereferencable (that
depends on the allocated size at runtime), such a global must be zero-
sized, which makes global variables a bad fit.

Group-shared memory in amdgpu and nvptx lives in address space 3.
Dynamic group-shared memory is implemented by creating an external
global variable in address space 3. The global is declared with size 0,
as the actual size is only known at runtime. It is defined behavior in
LLVM to access an external global outside the defined size.

There is no similar way to get the allocated size of dynamic shared
memory on amdgpu an nvptx, so users have to pass this out-of-band or
rely on target specific ways.